EDSC-2669: Rename action and reducer files.
